+++ src/metatags/handler.js
@@ -34,7 +34,8 @@ async function fetchAndProcessPageObject(s3Client, bucketName, key, prefix, log)
 }
 
 export default async function auditMetaTags(message, context) {
-  const { type, url: siteId } = message;
+  const { type } = message;
+  const siteId = message.siteId || message.url;
   const {
     dataAccess, log, s3Client,
   } = context;